tests/wpt/tests/html/browsers/browsing-the-web/history-traversal/pageswap/pageswap-cross-origin.sub.html

<!DOCTYPE HTML>
<title>Tests pageswap for cross-origin navigations</title>
<link rel="author" title="Khushal Sagar"  href="mailto:khushalsagar@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
const expectedUrl = get_host_info().HTTPS_REMOTE_ORIGIN + "/html/browsers/browsing-the-web/history-traversal/pageswap/pageswap-cross-origin.sub.html?new";

const params = new URLSearchParams(location.search);
// The page the popup navigates to.
const is_new_page = params.has('new');
// The initial page in the popup.
const is_popup_page = params.has('popup');
// The test page itself.
const is_test_page = !is_popup_page && !is_new_page;

const channel = new BroadcastChannel("testchannel");

if (is_test_page) {
  const expectedUrl = location.href + "?new";
  const expectedEvents = ["pageswap", "pagehide"];

  promise_test(async t => {
    let popup;
    onload = () => {
      window.events = [];
      popup = window.open("?popup");
    };

    await new Promise(resolve => {
      channel.addEventListener(
        "message", t.step_func(async (e) => {
          if (e.data === "nav") {
            assert_array_equals(window.events, expectedEvents, 'incorrect event order');
            popup.close();
            resolve();
          }
      }));
    });
  }, `pageswap on navigation from script`);
} else if (is_popup_page) {
  onload = () => {
    requestAnimationFrame(() => requestAnimationFrame(() => {
      location.href = expectedUrl;
    }));

    onpageswap = (e) => {
      window.opener.events.push("pageswap");
      if (e.activation != null)
        window.opener.events.push("activation");
      if (e.viewTransition != null)
        window.opener.events.push("transition");
    };

    onpagehide = () => {
      window.opener.events.push("pagehide");
      channel.postMessage("nav");
    };
  };
}
</script>
